requirements.txt
langchain==0.2.16
langchain-community==0.2.16
langchain-google-genai==1.0.10
langserve[all]==0.2.3
langchain-chroma==0.1.4
google-generativeai==0.7.2
chromadb==0.5.3
PyMuPDF==1.24.10
pdfplumber==0.11.4
python-docx==1.1.2
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
gradio==4.42.0
sentence-transformers==3.1.1
python-dotenv==1.0.1
httpx==0.27.2
pydantic==2.9.2
tiktoken==0.7.0
ragas==0.1.21